rewrited update and delete method. added search functionality

+++ src/main/java/de/odisho/store/controller/ProductController.java
@@ -1,5 +1,6 @@
 package de.odisho.store.controller;
 
+import java.io.IOException;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,7 +13,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestPart;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,6 +20,7 @@ import org.springframework.web.multipart.MultipartFile;
 
 import de.odisho.store.modle.Product;
 import de.odisho.store.service.ProductService;
+import org.springframework.web.bind.annotation.RequestParam;
 
 @RestController
 @CrossOrigin
@@ -53,7 +54,6 @@ public class ProductController {
         } catch (Exception e) {
             return new ResponseEntity<>(e.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
         }
-
     }
 
     @GetMapping("/product/{productId}/image")
@@ -64,13 +64,39 @@ public class ProductController {
         return ResponseEntity.ok().contentType(MediaType.valueOf(product.getImageType())).body(imageFile);
     }
 
-    @PutMapping("/products")
-    public void updateProduct(@RequestBody Product product) {
-        service.updateProduct(product);
+    @PutMapping("product/{id}")
+    public ResponseEntity<String> updateProduct(@PathVariable int id, @RequestPart Product product,
+            @RequestPart MultipartFile imageFile) {
+        Product prod;
+        try {
+            prod = service.updateProduct(id, product, imageFile);
+        } catch (IOException e) {
+            e.printStackTrace();
+            return new ResponseEntity<>("Faile to update", HttpStatus.BAD_REQUEST);
+        }
+
+        if (prod != null) {
+            return new ResponseEntity<>("Updated", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Faile to update", HttpStatus.BAD_REQUEST);
+        }
     }
 
-    @DeleteMapping("/products/{id}")
-    public void deleteProductById(@PathVariable int id) {
-        service.deleteProductById(id);
+    @DeleteMapping("/product/{id}")
+    public ResponseEntity<String> deleteProductById(@PathVariable int id) {
+        Product product = service.getProductById(id);
+        if (product != null) {
+            service.deleteProductById(id);
+            return new ResponseEntity<>("Deleted", HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>("Product not found", HttpStatus.NOT_FOUND);
+        }
+    }
+
+    @GetMapping("/products/search")
+    public ResponseEntity<List<Product>> searchProduct(@RequestParam String keyword) {
+        List<Product> products = service.searchProducts(keyword);
+        System.out.println(keyword);
+        return new ResponseEntity<>(products, HttpStatus.OK);
     }
 }
